rename node package to match the path name
