+++ staging/src/k8s.io/cloud-provider/controllers/node/node_controller_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cloud
+package node
 
 import (
 	"context"
